requirements.txt
aniso8601==6.0.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
Click==7.0
ecdsa==0.18.0
exceptiongroup==1.1.1
Flask==1.0.3
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
idna==3.4
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2-binary==2.8.2
pyasn1==0.5.0
pycodestyle==2.10.0
python-jose==3.3.0
pytz==2019.1
requests==2.30.0
rsa==4.9
six==1.12.0
SQLAlchemy==1.3.4
tomli==2.0.1
urllib3==2.0.2
Werkzeug==0.15.5
WTForms==3.0.1
